Переписал, используя BigDecimal. Добавил тест корзины

+++ api/src/main/java/ru/kolobkevic/market/api/dtos/CartDto.java
@@ -2,14 +2,15 @@ package ru.kolobkevic.market.api.dtos;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class CartDto {
     private List<CartItemDto> cartItemList;
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 
-    public CartDto(List<CartItemDto> items, Double totalPrice) {
+    public CartDto(List<CartItemDto> items, BigDecimal totalPrice) {
         this.cartItemList = items;
         this.totalPrice = totalPrice;
     }
